requirements.txt
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
imageio==2.31.1
lazy_loader==0.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.3
pyreadline3==3.4.1
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
scikit-image==0.21.0
scipy==1.11.1
shiboken6==6.5.1.1
six==1.16.0
sympy==1.12
tifffile==2023.4.12
typing_extensions==4.7.0
tzdata==2023.3
